+++ Widgets/FarmingBar_LuaEditor.lua
@@ -16,11 +16,22 @@ local Version = 1
 local methods = {
     OnAcquire = function(self)
         self.frame:Hide()
+
+        local editbox = self.editBox
+        addon.indent.enable(editbox, _, 4) -- adds syntax highlighting
+        self:SetUserData("OnUpdate", editbox:GetScript("OnUpdate"))
+        editbox:SetScript("OnUpdate", function(this)
+            self:GetUserData("OnUpdate")(this)
+            this.obj.button:Enable()
+        end)
     end,
 
     OnRelease = function(self)
+        local editbox = self.editBox
+        addon.indent.disable(self.editBox) 
+        editbox:SetScript("OnUpdate", self:GetUserData("OnUpdate"))        
+
         self.window.obj:Release()
-        addon.indent.disable(self.editbox.editBox)
         self.editbox.obj:Release()
     end,
 
@@ -53,11 +64,16 @@ local function Constructor()
     frame:Show()
 
     local editbox = AceGUI:Create("MultiLineEditBox")
-    addon.indent.enable(editbox.editBox, _, 4) -- adds syntax highlighting
     editbox:SetLabel("")
     window:AddChild(editbox)
 
-    editbox:SetCallback("OnEnterPressed", function(self, _, text)
+    editbox:SetCallback("OnEnterPressed", function(self, _, text)                                                                
+        local success, err = pcall(loadstring("return " .. text))
+        if not success then
+            window:SetStatusText(L["Error"] .. ": " .. err)
+            return
+        end
+
         local info = self:GetUserData("info")
         if info[5] then
             addon:SetBarDBValue(info[2], text, info[5])
@@ -78,15 +94,7 @@ local function Constructor()
 
         -- Update preview while typing
         local preview, err = func(addon, addon.unpack(args, {}), self:GetText())
-        window:SetStatusText(preview)
-
-        C_Timer.After(0.01, function()
-            if err then
-                editbox.button:Disable()
-            else
-                editbox.button:Enable()
-            end
-        end)
+        window:SetStatusText(preview or err)
     end)
 
     local widget = {
@@ -94,6 +102,7 @@ local function Constructor()
         window = window,
         frame = frame,
         editbox = editbox,
+        editBox = editbox.editBox,
         statustext = window.statustext,
     }