+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - You can now open the objective editor from an empty button to link an objective template
 - Implemented auto loot on use
 - Implemented keybind configurations
+- Implemented bar custom hide function
 - Bug fixes:
   - -- Fixed bug where bar was movable with the wrong button click
   - -- Fixed lua error when configuring bars after changing profiles